chore: refact task definition
