+++ crates/primitives/src/datalake/mod.rs
@@ -7,6 +7,7 @@ pub mod block_sampled;
 pub mod datalake_type;
 pub mod envelope;
 pub mod output;
+pub mod task;
 pub mod transactions;
 
 pub trait DatalakeCollection {
